# openapi.yaml
openapi: 3.1.0
info:
  title: 企业短信管理平台 API
  version: 1.6.0
  description: |
    统一短信发送与管理平台对外接口。
    - /api/v1/messages/** 供内部系统调用，X-Api-Key 认证（双 Key 轮换宽限期内新旧均有效）
    - /api/v1/web/** 供 Vue3 前端调用，显式选择本地或 AD 认证源；15 分钟 access JWT 每次按数据库权威安全上下文校验，refresh JWT 七天内单次轮换
    - 所有时间均为 ISO8601 +08:00
    - v1.1：消息分类(category)、biz_id 幂等、失败重发、结果回调、双 Key 轮换、登出
    - v1.2：计费条(est_segments/quota_cost)、模板变量渲染(template_params)、号码级频控(removed_freq_limit)、
      测试发送(is_test)、营销同意(consent_confirmed)、批次列表/号码搜索/队列恢复/用户管理/回调密钥轮换
    - v1.3-v1.5：厂商{sN}模板转换、OTP打码、unmatched、异常检测、号码时间线、任务健康
    - v1.6：AUTH_MOCK、完整CRUD/计费预估/改期/授权解密、密文导出、owner/app角色与确定性Mock验收
    - v1.6.13：用户角色变更与强制下线使用数据库会话版本即时失效既有 JWT
    - v1.6.14：本地/AD Provider 显式路由、稳定主体 JWT 与首次强制改密
    - v1.6.24：统一 security_version 覆盖账号/身份/Provider/部门/角色，短 access JWT 与单次 refresh 轮换
    - v1.6.12：运行策略事务最终校验、动态过期时间与启动调度心跳一致
    - v1.6.11：报告回调引用新增无 PII 事件身份，raw 重放不重复投递同一报告
    - v1.6.28：callback/export 使用可续租 UUID fencing；回调暴露稳定 event_id 与租约状态
    - v1.6.33：拆分无依赖 livez 与严格依赖 readyz；健康响应不暴露内部组件或失败原因
    - v1.6.34：生产 callback 强制 HTTPS；签名密钥/协议版本按事件固化；敏感密文与导出帧绑定业务上下文
servers:
  - url: https://sms.example.com
...
paths:
...
  /api/v1/web/replies:
    get:
      tags: [replies]
      summary: 上行回复查询（手机号精确查询走 HMAC 索引）
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: phone, in: query, required: false, schema: { anyOf: [{ type: string, pattern: '^1\d{10}$' }, { type: 'null' }] } }
        - { name: start, in: query, required: false, schema: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] } }
        - { name: end, in: query, required: false, schema: { anyOf: [{ type: string, format: date-time }, { type: 'null' }] } }
        - { name: page, in: query, required: false, schema: { type: integer, minimum: 1, default: 1 } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [total, items]
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Reply' }
        '400': { description: INVALID_PARAM, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/web/replies/{id}/blacklist:
    post:
      tags: [replies]
      summary: 将上行回复号码以 reply_optout 来源加入黑名单
      security: [{ BearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: 已加入或幂等命中现有黑名单 }
        '403': { description: FORBIDDEN, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────── 管理 ─────────────────────────
...
components:
  securitySchemes:
    MetricsBearerAuth:
      type: http
      scheme: bearer
      description: 仅供 Prometheus 抓取使用的独立、可轮换 secret。
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: 资源不存在
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
...
    Error:
      type: object
      required: [code, message, detail]
      properties:
        code: { type: string, description: 见 CLAUDE.md 平台错误码表 }
        message: { type: string }
        detail: { type: object, nullable: true }
...
    Reply:
      type: object
      required: [id, phone, content, batch_no, reply_time]
      properties:
        id: { type: integer }
        phone: { type: string, description: 掩码展示 }
        content: { type: string }
        batch_no: { anyOf: [{ type: string }, { type: 'null' }] }
        reply_time: { type: string, format: date-time }
...